PR comment suggested changes
