+++ auth/providers/azure/entrasdk_verifier_test.go
